refactor: move tanstack generator to separate file
