+++ packages/cli/src/generators/tanstack-react/writeOpenAPIServices.ts
@@ -5,30 +5,11 @@ import { URL } from 'node:url';
 import { OpenAPI3 } from 'openapi-typescript';
 import ora, { Ora } from 'ora';
 
-import {
-  getServices,
-  Service,
-  ServiceBaseName,
-} from '../../lib/open-api/getServices.js';
+import { getServices } from '../../lib/open-api/getServices.js';
 import { readSchema } from '../../lib/open-api/readSchema.js';
-import { astToString } from './ts-factory/astToString.js';
-import { getClientFactory } from './ts-factory/getClientFactory.js';
-import { getIndexFactory } from './ts-factory/getIndexFactory.js';
-import {
-  getServiceFactory,
-  ServiceImportsFactoryOptions,
-} from './ts-factory/getServiceFactory.js';
-import { getServiceIndexFactory } from './ts-factory/getServiceIndexFactory.js';
-
-type OutputOptions = {
-  fileHeader?: string;
-  dir: URL;
-  clean: boolean;
-  postfixServices?: string;
-  explicitImportExtensions?: boolean;
-  servicesDirName: string;
-  serviceNameBase?: ServiceBaseName;
-};
+import { OutputOptions } from '../../lib/OutputOptions.js';
+import { generateCode, GeneratorFiles, getFileHeader } from './generateCode.js';
+import { ServiceImportsFactoryOptions } from './ts-factory/getServiceFactory.js';
 
 export const writeOpenAPIServices = async ({
   source,
@@ -128,185 +109,4 @@ const writeFiles = async (
       throw error;
     }
   }
-};
-
-const generateCode = async ({
-  spinner,
-  services,
-  serviceImports,
-  output,
-}: {
-  spinner: Ora;
-  services: Service[];
-  serviceImports: ServiceImportsFactoryOptions;
-  output: OutputOptions;
-}) => {
-  return [
-    ...(await generateServices(spinner, services, serviceImports, output)),
-    ...(await generateServiceIndex(spinner, services, output)),
-    ...(await generateClient(spinner, output)),
-    ...(await generateIndex(spinner, output)),
-  ];
-};
-
-type GeneratorFiles = Array<
-  { file: URL; code: string } | { directory: URL; clean: boolean }
->;
-
-const composeServicesDirPath = (
-  output: Pick<OutputOptions, 'servicesDirName' | 'dir'>
-) => {
-  return new URL(`${output.servicesDirName}/`, output.dir);
-};
-
-const generateServices = async (
-  spinner: Ora,
-  services: Service[],
-  serviceImports: ServiceImportsFactoryOptions,
-  output: OutputOptions
-) => {
-  const servicesDir = composeServicesDirPath(output);
-
-  spinner.start('Generating services');
-
-  const serviceFiles: GeneratorFiles = [
-    {
-      directory: servicesDir,
-      clean: output.clean,
-    },
-  ];
-
-  for (const service of services) {
-    const { operations, name, typeName, variableName, fileBaseName } = service;
-
-    spinner.text = `Generating ${c.magenta(name)} service`;
-
-    try {
-      const code = astToString(
-        getServiceFactory(
-          {
-            typeName,
-            variableName,
-          },
-          operations,
-          serviceImports
-        )
-      );
-
-      serviceFiles.push({
-        file: new URL(`${fileBaseName}.ts`, servicesDir),
-        code,
-      });
-    } catch (error) {
-      spinner.fail(
-        c.redBright(
-          `Error occurred during ${c.magenta(name)} service generation`
-        )
-      );
-
-      throw error;
-    }
-  }
-
-  spinner.succeed(c.green('Services has been generated'));
-
-  return serviceFiles;
-};
-
-const generateServiceIndex = async (
-  spinner: Ora,
-  services: Service[],
-  output: OutputOptions
-) => {
-  spinner.start('Generating services index');
-
-  const serviceIndexFiles: GeneratorFiles = [
-    {
-      directory: composeServicesDirPath(output),
-      clean: output.clean,
-    },
-  ];
-
-  try {
-    const code = astToString(
-      getServiceIndexFactory(services, {
-        explicitImportExtensions: Boolean(output.explicitImportExtensions),
-      })
-    );
-
-    serviceIndexFiles.push({
-      file: new URL('index.ts', composeServicesDirPath(output)),
-      code,
-    });
-  } catch (error) {
-    spinner.fail(
-      c.redBright('Error occurred during services index generation')
-    );
-
-    throw error;
-  }
-
-  spinner.succeed(c.green('Services index has been generated'));
-
-  return serviceIndexFiles;
-};
-
-const generateClient = async (spinner: Ora, output: OutputOptions) => {
-  spinner.start('Generating client');
-
-  const clientFiles: GeneratorFiles = [];
-
-  try {
-    const code = astToString(
-      getClientFactory({
-        servicesDirName: output.servicesDirName,
-        explicitImportExtensions: Boolean(output.explicitImportExtensions),
-      })
-    );
-
-    clientFiles.push({
-      file: new URL('create-api-client.ts', output.dir),
-      code,
-    });
-  } catch (error) {
-    spinner.fail(c.redBright('Error occurred during client generation'));
-
-    throw error;
-  }
-
-  spinner.succeed(c.green('Client has been generated'));
-
-  return clientFiles;
-};
-
-const generateIndex = async (spinner: Ora, output: OutputOptions) => {
-  spinner.start('Generating index');
-
-  const indexFiles: GeneratorFiles = [];
-
-  try {
-    const code = astToString(
-      getIndexFactory({
-        servicesDirName: output.servicesDirName,
-        explicitImportExtensions: Boolean(output.explicitImportExtensions),
-      })
-    );
-
-    indexFiles.push({
-      file: new URL('index.ts', output.dir),
-      code,
-    });
-  } catch (error) {
-    spinner.fail(c.redBright('Error occurred during index generation'));
-
-    throw error;
-  }
-
-  spinner.succeed(c.green('Index has been generated'));
-
-  return indexFiles;
-};
-
-const getFileHeader = ({ fileHeader }: Pick<OutputOptions, 'fileHeader'>) => {
-  return fileHeader && `${fileHeader}\n`;
 };
