+++ packages/cli/src/lib/OutputOptions.ts
@@ -1,0 +1,13 @@
+import { URL } from 'node:url';
+
+import { ServiceBaseName } from './open-api/getServices.js';
+
+export type OutputOptions = {
+  fileHeader?: string;
+  dir: URL;
+  clean: boolean;
+  postfixServices?: string;
+  explicitImportExtensions?: boolean;
+  servicesDirName: string;
+  serviceNameBase?: ServiceBaseName;
+};
